requirements.txt
# core framework
streamlit

#langchain core
langchain
langchain-core
langchain-community

# google api integration
langchain-google-genai
google-generativeai

# utilities
Pillow
python-dotenv
pydantic<2